JERSEY-4654 JerseyQueryParamStyle support
