+++ core-client/src/main/java/org/glassfish/jersey/client/ClientProperties.java
@@ -451,6 +451,18 @@ public final class ClientProperties {
      */
     public static final Long DEFAULT_EXPECT_100_CONTINUE_THRESHOLD_SIZE = 65536L;
 
+    /**
+     * The property defines the desired format of query param when multiple
+     * values are sent for the same parameter.
+     *
+     * <p>
+     * The value MUST be an instance of
+     * {@link org.glassfish.jersey.uri.QueryParamStyle}.</p>
+     * <p>
+     * The default value is {@code null}.</p>
+     */
+    public static final String QUERY_PARAM_STYLE = "jersey.config.client.uri.query.param.style";
+
     private ClientProperties() {
         // prevents instantiation
     }
